requirements.txt
# ─── Phase 1: Core Valuation Engine ───────────────────────────────
pandas>=2.0
numpy>=1.24
yfinance>=0.2.18
python-dotenv>=1.0
requests>=2.28

# ─── Phase 2: Enhanced Data Pipeline ──────────────────────────────
# (requests already included above for FRED API)
lxml>=4.9           # For Wikipedia S&P 500 scraping
html5lib>=1.1       # Fallback HTML parser

# ─── Phase 3: AI Agent Layer ──────────────────────────────────────
groq>=0.4           # Groq API client (free tier)

# ─── Phase 4: ML Enhancement ─────────────────────────────────────
scikit-learn>=1.3
xgboost>=2.0
# lightgbm>=4.0     # Optional: faster alternative to XGBoost
# mapie>=0.8        # Optional: conformal prediction intervals

# ─── Phase 5: Web Dashboard ──────────────────────────────────────
fastapi>=0.100
uvicorn>=0.22

# ─── Testing ──────────────────────────────────────────────────────
pytest>=7.0
